Refine player jump and ground collision mechanics
The player's jump action in PlayerViewModel was simplified by removing an unnecessary condition, which previously checked the player's vertical speed before setting it to the predefined jump speed. In CollisionViewModel, the game's response to a ground collision was altered: now the entity is considered "on ground" when the shifted collision is approximately one unit inside the floor. The parameters of the respondToGroundCollision function were accordingly adjusted in CollisionViewModel.h to reflect these changes.